move feedbackurlconfig to base
