--- eahub/config/apps.py
@@ -1,6 +1,0 @@
-from django import apps
-
-
-class ConfigConfig(apps.AppConfig):
-    name = "eahub.config"
-    verbose_name = "Config"
